Remove "." from state abbreviations in address line, and fix bug where title uppercase did not get protected
Mote details:

Sentence case in the title line was only protected when the start of the sentence was lowercase, changed the regex to be case insensitive so that also correctly formatted uppercases get protected from the forced lowercasing in the bst script. 

Also changed file write mode to "w" instead of "a", though not that important.